serial_settings

Struct Runner

pub struct Runner<'a, P: Platform>(/* private fields */);

Implementations

impl<'a, P: Platform> Runner<'a, P>

pub fn new( platform: P, line_buf: &'a mut [u8], serialize_buf: &'a mut [u8], settings: &mut P::Settings, ) -> Result<Self, P::Error>

Constructor

Args

platform - The platform associated with the serial settings, providing the necessary context and API to manage device settings.

line_buf - A buffer used for maintaining the serial menu input line. It should be at least as long as the longest user input.

serialize_buf - A buffer used for serializing and deserializing settings. This buffer needs to be at least as big as twice the biggest serialized setting plus its path.

pub fn interface_mut(&mut self) -> &mut P::Interface

Get the device communication interface

pub fn platform_mut(&mut self) -> &mut P

pub fn platform(&mut self) -> &P

pub fn poll( &mut self, settings: &mut P::Settings, ) -> Result<bool, <P::Interface as ErrorType>::Error>

Must be called periodically to process user input.

Returns

A boolean indicating true if the settings were modified.